requirements.txt
flask
python-dotenv
pinecone
groq
requests
beautifulsoup4
sentence-transformers
flask-cors
